# sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProposalCreate": {
        "properties": {
          "category": {
            "default": "General",
            "maxLength": 60,
            "title": "Category",
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "pass_threshold": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Pass Threshold",
            "type": "integer"
          },
          "quorum_threshold": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quorum Threshold",
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "voting_period_days": {
            "default": 7,
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Voting Period Days",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "description"
        ],
        "title": "ProposalCreate",
        "type": "object"
      },
      "ProposalDetailRead": {
        "properties": {
          "abstain_pct": {
            "title": "Abstain Pct",
            "type": "number"
          },
          "against_pct": {
            "title": "Against Pct",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "executed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "executed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed By"
          },
          "for_pct": {
            "title": "For Pct",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "pass_threshold": {
            "title": "Pass Threshold",
            "type": "integer"
          },
          "proposer_address": {
            "title": "Proposer Address",
            "type": "string"
          },
          "quorum_met": {
            "title": "Quorum Met",
            "type": "boolean"
          },
          "quorum_threshold": {
            "title": "Quorum Threshold",
            "type": "integer"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_abstain": {
            "title": "Total Abstain",
            "type": "integer"
          },
          "total_against": {
            "title": "Total Against",
            "type": "integer"
          },
          "total_for": {
            "title": "Total For",
            "type": "integer"
          },
          "turnout_pct": {
            "title": "Turnout Pct",
            "type": "number"
          },
          "user_vote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoteChoice"
              },
              {
                "type": "null"
              }
            ]
          },
          "votes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VoteRead"
            },
            "title": "Votes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "category",
          "proposer_address",
          "status",
          "start_time",
          "end_time",
          "quorum_threshold",
          "pass_threshold",
          "total_for",
          "total_against",
          "total_abstain",
          "created_at",
          "turnout_pct",
          "for_pct",
          "against_pct",
          "abstain_pct",
          "quorum_met"
        ],
        "title": "ProposalDetailRead",
        "type": "object"
      },
      "ProposalRead": {
        "properties": {
          "abstain_pct": {
            "title": "Abstain Pct",
            "type": "number"
          },
          "against_pct": {
            "title": "Against Pct",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "executed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "executed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed By"
          },
          "for_pct": {
            "title": "For Pct",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "pass_threshold": {
            "title": "Pass Threshold",
            "type": "integer"
          },
          "proposer_address": {
            "title": "Proposer Address",
            "type": "string"
          },
          "quorum_met": {
            "title": "Quorum Met",
            "type": "boolean"
          },
          "quorum_threshold": {
            "title": "Quorum Threshold",
            "type": "integer"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_abstain": {
            "title": "Total Abstain",
            "type": "integer"
          },
          "total_against": {
            "title": "Total Against",
            "type": "integer"
          },
          "total_for": {
            "title": "Total For",
            "type": "integer"
          },
          "turnout_pct": {
            "title": "Turnout Pct",
            "type": "number"
          },
          "user_vote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoteChoice"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "category",
          "proposer_address",
          "status",
          "start_time",
          "end_time",
          "quorum_threshold",
          "pass_threshold",
          "total_for",
          "total_against",
          "total_abstain",
          "created_at",
          "turnout_pct",
          "for_pct",
          "against_pct",
          "abstain_pct",
          "quorum_met"
        ],
        "title": "ProposalRead",
        "type": "object"
      },
      "ProposalStatus": {
        "description": "A governance proposal's lifecycle. finalize() (routers/governance.py)\nis the only thing that moves ACTIVE -> PASSED/REJECTED; EXECUTED is\nreserved for a future action that actually applies a passed proposal's\neffect (see ROADMAP.md Part 1) and isn't set by anything yet.",
        "enum": [
          "active",
          "passed",
          "rejected",
          "executed"
        ],
        "title": "ProposalStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoteChoice": {
        "enum": [
          "for",
          "against",
          "abstain"
        ],
        "title": "VoteChoice",
        "type": "string"
      },
      "VoteCreate": {
        "properties": {
          "choice": {
            "$ref": "#/components/schemas/VoteChoice"
          }
        },
        "required": [
          "choice"
        ],
        "title": "VoteCreate",
        "type": "object"
      },
      "VoteRead": {
        "properties": {
          "choice": {
            "$ref": "#/components/schemas/VoteChoice"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "proposal_id": {
            "title": "Proposal Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "voter_address": {
            "title": "Voter Address",
            "type": "string"
          },
          "voting_power": {
            "title": "Voting Power",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "proposal_id",
          "voter_address",
          "choice",
          "voting_power",
          "created_at",
          "updated_at"
        ],
        "title": "VoteRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for Nuance \u2014 AI-validator consensus for claims traditional smart contracts can't settle.",
    "title": "Nuance API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/proposals": {
      "get": {
        "operationId": "list_proposals_proposals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalRead"
                  },
                  "title": "Response List Proposals Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Proposals",
        "tags": [
          "governance"
        ]
      },
      "post": {
        "operationId": "create_proposal_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Proposal",
        "tags": [
          "governance"
        ]
      }
    },
    "/proposals/{proposal_id}": {
      "get": {
        "operationId": "get_proposal_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Proposal",
        "tags": [
          "governance"
        ]
      }
    },
    "/proposals/{proposal_id}/execute": {
      "post": {
        "description": "Moves a PASSED proposal to EXECUTED \u2014 the formal \"this decision has\nbeen enacted\" marker. Not idempotent like finalize: a proposal can only\nbe executed once, so a second call 400s rather than silently returning\nthe same result, matching disputes.py's enforce_ruling precedent for\nthe same \"who did this and when\" shape (executed_by/executed_at here,\nenforced_by/resolved_at there).\n\nNo real on-chain effect is wired up yet (no treasury transfer, no\nparameter change) \u2014 see ROADMAP.md Part 3; this is deliberately scoped\nto just the status transition until there's a real effect to apply.",
        "operationId": "execute_proposal_proposals__proposal_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Execute Proposal",
        "tags": [
          "governance"
        ]
      }
    },
    "/proposals/{proposal_id}/finalize": {
      "post": {
        "description": "Idempotent \u2014 an already-finalized proposal is returned as-is rather\nthan re-evaluated, so calling this twice is always safe. Raises 400 if\n`end_time` hasn't arrived yet; a future bulk sweep (ROADMAP.md Part 1's\ncron) should filter to `end_time <= now()` itself before calling this\nper-id rather than relying on the 400 to skip early proposals.",
        "operationId": "finalize_proposal_proposals__proposal_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Proposal",
        "tags": [
          "governance"
        ]
      }
    },
    "/proposals/{proposal_id}/vote": {
      "post": {
        "operationId": "cast_vote_proposals__proposal_id__vote_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cast Vote",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}